prompt_tracking_extension/requirements.txt
flask
gunicorn
os
pandas
datetime
textblob
google.generativeai
dotenv
transformers
stanza
requests